+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog for NVDB API LES V3 Client
 
+## 1.3.0
+* Fixed bug when parsing RoadSysRef of road segments in object.
+* Fixed bug when formatting TimeAttribute.
+* Section.direction, Intersection.direction, SideArea.direction changed from 
+String to Direction.
+* RoadSystem.roadCategory, RoadSystem.phase changed from String to enum.
 
 ## 1.2.2
 * Fixed NPE when Roadsystem does not have number.